# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/payments": {
      "get": {
        "operationId": "PaymentController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PaymentResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all payments",
        "tags": [
          "payments"
        ]
      },
      "post": {
        "operationId": "PaymentController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a payment",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/payments/paypal-create-order": {
      "post": {
        "operationId": "PaymentController_createPaypalOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaypalOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a PayPal order",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/payments/paypal-capture-order": {
      "post": {
        "operationId": "PaymentController_captureOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "orderId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Capture a PayPal order",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/payments/apple-validate-merchant": {
      "post": {
        "operationId": "PaymentController_appleValidateMerchant",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "validationUrl": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Validate Apple Pay merchant",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/payments/{id}": {
      "get": {
        "operationId": "PaymentController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get a payment by ID",
        "tags": [
          "payments"
        ]
      },
      "patch": {
        "operationId": "PaymentController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponseDto"
                }
              }
            }
          }
        },
        "summary": "Update a payment",
        "tags": [
          "payments"
        ]
      }
    }
  },
  "info": {
    "title": "Stansbury Swim API",
    "description": "API documentation for Stansbury Swim backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PaymentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "paymentGateway": {
            "type": "object"
          },
          "paymentGatewayId": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "quantity": {
            "type": "number"
          },
          "status": {
            "type": "object"
          },
          "scheduleId": {
            "type": "string"
          },
          "studentId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "userId",
          "productId",
          "paymentGateway",
          "paymentGatewayId",
          "amount",
          "quantity",
          "status"
        ]
      },
      "CreatePaypalOrderDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "scheduleId": {
            "type": "string"
          },
          "studentId": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "productId",
          "quantity"
        ]
      },
      "CreatePaymentDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "paymentGateway": {
            "type": "string",
            "enum": [
              "PAYPAL",
              "APPLE_PAY"
            ]
          },
          "paymentGatewayId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "SUCCESS",
              "FAILED"
            ],
            "default": "PENDING"
          },
          "scheduleId": {
            "type": "string"
          },
          "studentId": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "productId",
          "quantity",
          "paymentGateway",
          "paymentGatewayId",
          "status"
        ]
      },
      "UpdatePaymentDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}